Hack to change source number counting for for_objects
For the coord_file approach for Twinkles, there was one coord_file
for all visits.  Thus the photometry ID was tracked by which row
in the coord file.

But for the for_objects approach, there is one coord file for each
of many fields.  Here, the object is always the first.  Or, in more generality,
separate accounting should be done for each entry in the source table
for each field.